Moved repository into Visual Studio project
